flask/requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
aniso8601==9.0.1
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
beautifulsoup4==4.12.0
bs4==0.0.1
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
chromedriver==2.24.1
click==8.1.3
colorama==0.4.6
cssselect==1.2.0
decorator==4.4.2
distlib==0.3.6
dnspython==2.3.0
exceptiongroup==1.1.1
feedfinder2==0.0.4
feedparser==6.0.10
filelock==3.10.7
Flask==2.2.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
Flask-SQLAlchemy==3.0.3
frozenlist==1.3.3
greenlet==2.0.2
h11==0.14.0
html5lib==1.1
idna==3.4
imageio==2.27.0
imageio-ffmpeg==0.4.8
itsdangerous==2.1.2
jieba3k==0.35.1
Jinja2==3.1.2
joblib==1.2.0
lxml==4.9.2
MarkupSafe==2.1.2
moviepy==1.0.3
multidict==6.0.4
newspaper3k==0.2.8
nltk==3.8.1
numpy==1.24.2
oauthlib==3.2.2
openai==0.27.4
outcome==1.2.0
Pillow==9.5.0
platformdirs==3.2.0
proglog==0.1.10
pycodestyle==2.10.0
pycparser==2.21
pymongo==4.3.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2023.3
pywin32==306
PyYAML==6.0
regex==2023.3.23
requests==2.28.2
requests-file==1.5.1
requests-oauthlib==1.3.1
selenium==4.8.3
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
SpeechRecognition==3.10.0
SQLAlchemy==2.0.8
tinysegmenter==0.3
tldextract==3.4.0
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
tweepy==4.13.0
typing_extensions==4.5.0
urllib3==1.26.15
virtualenv==20.21.0
webencodings==0.5.1
Werkzeug==2.2.3
wsproto==1.2.0
yarl==1.8.2
